Java Docs and JUnit

+++ oral_exam1/S27_MachineLearning_Hard/src/Euclidean.java
@@ -1,8 +1,8 @@
 import java.lang.Math;
 
 public class Euclidean {
-    public static double euclideanDistance(int[] num1, int[] num2) {
-        int temp = 0;
+    public static double euclideanDistance(double[] num1, double[] num2) {
+        double temp = 0;
 
         if (num1.length == num2.length) {
             for (int i = 0; i < num1.length; i++) {
